% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{aids_brasil}
\alias{aids_brasil}
\title{Monthly notified cases of AIDS in Brazil}
\format{
A \code{\link{data.frame}} with 40 observations and 2 columns:
\itemize{
\item \code{time}: Date. The correspond month.
\item \code{cases}: Numeric. Cumulative monthly notifications of AID cases.
}
}
\source{
Gamerman, D. and Migon, H. S. (1991). Forecasting the number of AIDS cases in Brasil,  \emph{The Statistician}, \bold{40}, 427--442.
}
\usage{
data(aids_brasil, package = "RBATS")
}
\description{
Data from Gamerman and Migon (1991).
It contains the cumulative number of monthly notifications of AID cases in Brazil
from September 1985 to December 1988.
}
\author{
André Felipe Menezes
}
\keyword{datasets}
