% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{association_index}
\alias{association_index}
\title{Calculate AI}
\usage{
association_index(thisnode)
}
\arguments{
\item{node}{the node whose AI should be calculated}
}
\value{
the association index
}
\description{
Calculate the association index of one single node
}
