% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.dlm.fit}
\alias{autoplot.dlm.fit}
\title{Fit plot method for Bayesian Dynamic Linear Models.}
\usage{
\method{autoplot}{dlm.fit}(
  object,
  what = c("predictive", "posterior"),
  type = c("filter", "smooth", "both"),
  interval = type == "filter",
  ...
)
}
\arguments{
\item{object}{fitted model object of class \code{dlm.fit}.}

\item{what}{character. Should plot the \code{"predictive"} or \code{"posterior"}
state moments?. Default is \code{"predictive"}.}

\item{type}{character. Should plot the \code{"filter"}, \code{"smooth"}, or
\code{"both"}?. Default is \code{"filter"}.}

\item{interval}{logical. Should exact credible interval be plotted? Default is
\code{TRUE} for filter distribution and FALSE for \code{"smooth"} and \code{"both"}.}

\item{...}{additional plotting arguments that affect the plot.}
}
\description{
Method \code{autoplot} applied to objects of class \code{dlm.fit}.
Plot historical data with forecasts and credible interval. Also, plot the filter and
smooth state parameters.
}
\author{
André F. B. Menezes
}
