% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{bats}
\alias{bats}
\title{Analyse a single tree file}
\usage{
bats(treefile, xmlfile, reps = 1, userinput = NULL)
}
\arguments{
\item{treefile}{the path to the tree file}

\item{xmlfile}{the path to the .xml file}

\item{reps}{how many shuffled trees should be made for each normal tree}

\item{userinput}{the state attribute}
}
\value{
a data frame containing the output statistics
}
\description{
Calculate all statistics of a tree file
}
