% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{calculate_AI}
\alias{calculate_AI}
\title{Calculate AI}
\usage{
calculate_AI(node)
}
\arguments{
\item{node}{the tree whose AI should be calculated}
}
\value{
the association index
}
\description{
Calculate the association index of the whole tree
}
