% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_all_stats}
\alias{calculate_all_stats}
\title{Calculate all the statistics}
\usage{
calculate_all_stats(
  treelist,
  utree,
  possible_states,
  state_attribute,
  xmldict,
  all_stats
)
}
\arguments{
\item{treelist}{the tree}

\item{utree}{the raw tree}

\item{possible_states}{a vector of all possible states}

\item{state_attribute}{the state attribute}

\item{xmldict}{the xml dictionary}

\item{all_stats}{the list containing all the stats of all previous trees}
}
\value{
the list containing all the stats of all previous trees with the stats of this tree
}
\description{
Calculate all the statistics for the given tree and add them to the list of calculated stats
}
