% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{count_leaves}
\alias{count_leaves}
\title{Get leaf count}
\usage{
count_leaves(node)
}
\arguments{
\item{node}{the node whose leaves should counted}
}
\value{
the total amount of leaves
}
\description{
Count all the leaves under this node
}
