% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cp6}
\alias{cp6}
\title{Monthly total sales of tobacco from a company in UK.}
\format{
A \code{\link{data.frame}} with 60 observations and 2 columns.
}
\source{
XXXX.
}
\usage{
data(cp6, package = "RBATS")
}
\description{
Monthly total sales, in monetary terms on a standard scale, of
tobacco and related products marketed by a major company in the UK.
The time of the data runs from January 1955 to December 1959 inclusive.
}
\author{
André Felipe Menezes
}
\keyword{datasets}
