% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgegm.R
\name{dgegm}
\alias{dgegm}
\title{Bayesian Dynamic Generalized Exponential Growth Model}
\usage{
dgegm(
  lambda = 1,
  discount_factors,
  df_variance = 1,
  variance_law = list(type = "identity", p = 2),
  prior = NULL,
  n = 1,
  s = 1
)
}
\arguments{
\item{lambda}{numeric. Box-Cox power parameter \eqn{\lambda}. Special cases are:
\eqn{\lambda=1} the Modified Exponential curve, \eqn{\lambda=-1} the Logistic curve,
and \eqn{\lambda=0} the Gompertz curve.}

\item{discount_factors}{list. Discount factors for each model parameters,
\eqn{\boldsymbol{\theta}_t = (\theta_{1t}, \theta_{2t}, \theta_{3t})}.}

\item{df_variance}{Discount factor for observation variance. Use a beta-gamma random walk.}

\item{variance_law}{list. Variance law \code{type} and \code{power} parameter.
The variance law \code{type} are \code{identity}, \code{poisson}, \code{binomial},
and \code{power}. The variance law \code{power} should be numeric \eqn{p \geq 1}.}

\item{prior}{list. If specified it should contains prior for state parameters.}

\item{n, s}{numeric. Prior sample size and mean for the variance, respectively.
For both parameter the default is 1.}
}
\description{
Create object of class \code{dgegm}.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.

Migon, H. M. and Gamerman, D. (1993). Generalized exponential growth models: A Bayesian approach,  \emph{Journal of Forecasting}, \bold{12}, 573--584

Gamerman, D. and Migon, H. S. (1991). Forecasting the number of AIDS cases in Brasil,  \emph{The Statistician}, \bold{40}, 427--442.
}
\author{
André F. B. Menezes
}
