% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{dlm}
\alias{dlm}
\title{Bayesian Dynamic Linear Model}
\usage{
dlm(
  polynomial = list(order = 1L, discount_factor = 0.95),
  seasonal = list(type = c("none", "free", "fourier"), period = NULL, harmonics = NULL,
    discount_factor = 0.98),
  regressor = list(xreg = NULL, discount_factor = 0.99),
  autoregressive = list(order = NULL, discount_factor = 0.998),
  cycle = list(frequency = NULL, rho = NULL, discount_factor = 0.998),
  transfer_function = list(order = NULL, xreg = NULL, lambda = NULL, discount_factor =
    0.998),
  df_variance = 1
)
}
\arguments{
\item{polynomial}{list. Order of polynomial model.}

\item{seasonal}{list. Components to specify seasonal model.}

\item{regressor}{list. Components to specify a dynamic regression model.}

\item{autoregressive}{list. Components to specify a dynamic autoregressive model.}

\item{cycle}{list. Components to specify a dynamic cycle model.}

\item{transfer_function}{list. Components to specify a dynamic transfer function model.}

\item{df_variance}{numeric. Discount factor for observational variance. It uses a beta-gamma random walk.}
}
\description{
Create object of class \code{dlm}.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\author{
André F. B. Menezes
}
