% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dlm.R
\name{extract.dlm.fit}
\alias{extract.dlm.fit}
\alias{extract}
\title{Extract information of filtering and smoothing distribution from
\code{dlm.fit} object}
\usage{
extract(x, ...)

\method{extract}{dlm.fit}(
  x,
  prob_interval = c(0.05, 0.2),
  distribution = c("filter", "smooth"),
  component = c("response", "state"),
  ...
)
}
\arguments{
\item{x}{an object of class \code{\link{dlm.fit}}.}

\item{...}{currently not used.}

\item{prob_interval}{vector with the probability of the credibility interval.}

\item{distribution}{character. The options are  \code{filter} or \code{smooth}. Default is \code{filter}}

\item{which}{character. The options are \code{state} or \code{response}.}
}
\value{
A \code{data.frame} with the results for each time.
}
\description{
Tidy the results of the forward filter or backward smoother into
a \code{data.frame}.
}
