% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{finish_node}
\alias{finish_node}
\title{Finish internal node}
\usage{
finish_node(thisnode)
}
\arguments{
\item{thisnode}{the internal node that should be finished}
}
\value{
the finished internal node
}
\description{
Finishes the internal node by setting the states, state weights, and whether or not the node is monophyletic.
}
