% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit.dgegm}
\alias{fit.dgegm}
\alias{forward_filter.dgegm}
\title{Fitting Bayesian Dynamic Generalized Exponential Growth Model}
\usage{
\method{fit}{dgegm}(
  object,
  y,
  smooth = FALSE,
  a0 = NULL,
  R0 = NULL,
  n = 1,
  s = 1,
  prior_length = 10L,
  interval = TRUE,
  level = 0.1,
  ...
)

\method{forward_filter}{dgegm}(
  object,
  y,
  a0 = NULL,
  R0 = NULL,
  n = 1,
  s = 1,
  prior_length = 10L,
  interval = TRUE,
  level = 0.1,
  ...
)
}
\arguments{
\item{object}{a model object of class \code{dgem}.}

\item{y}{vector. Observed value of time series.}

\item{smooth}{logical. Should smoothing be performed? Default is \code{TRUE}.}

\item{a0, R0}{vector and matrix. prior state mean vector and covariance matrix.
Default is \code{NULL}, then use the first \code{prior_length} observation to compute.}

\item{n, s}{numeric. Prior sample size and mean for the variance, respectively.
For both parameter the default is 1.}

\item{prior_length}{integer. Number of observations to use to obtain the prior moments.
Default is 10.}

\item{interval}{logical. Should exact credible interval be returned? Default is \code{TRUE}.}

\item{level}{numeric. The probability level of the credible interval. Default is \code{0.10}.}

\item{...}{currently not used.}

\item{state_parameters}{list. Historical prior moments (\code{a} and \code{R}) and
historical posterior moments (\code{m} and \code{C})}
}
\description{
Method \code{fit} applied to objects of class \code{dgem}
to perform the filtering and smoothing.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.

Migon, H. M. and Gamerman, D. (1993). Generalized exponential growth models: A Bayesian approach,  \emph{Journal of Forecasting}, \bold{12}, 573--584
}
\author{
André F. B. Menezes
}
