% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit.dlm}
\alias{fit.dgegm}
\title{Fitting Bayesian DLM and DGEGM}
\usage{
fit(object, y, ...)

\method{fit}{dlm}(
  object,
  y,
  smooth = TRUE,
  prior_length = 10L,
  interval = TRUE,
  level = 0.1,
  ...
)

\method{fit}{dgegm}(object, y, interval = TRUE, level = 0.1, ...)
}
\arguments{
\item{object}{a model object of class \code{dlm} or \code{dgem}.}

\item{y}{vector. Observed value of time series.}

\item{...}{currently not used.}

\item{smooth}{logical. Should smoothing be performed? Default is \code{TRUE}.}

\item{prior_length}{integer. Number of observations to use to obtain the prior moments.
Default is 10.}

\item{interval}{logical. Should exact credible interval be returned? Default is \code{TRUE}.}

\item{level}{numeric. The probability level of the credible interval. Default is \code{0.10}.}
}
\description{
Method \code{fit} applied to objects of class \code{dlm} or \code{dgem}
to perform the filtering and smoothing.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.

Prado, R.; West, M. Time Series Modeling, Computation, and Inference. CRC Press, 2010.

Migon, H. M. and Gamerman, D. (1993). Generalized exponential growth models: A Bayesian approach,  \emph{Journal of Forecasting}, \bold{12}, 573--584
}
\author{
André F. B. Menezes
}
