% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_dlm.R
\name{extract.dlm.fit}
\alias{extract.dlm.fit}
\alias{forecast}
\title{Extract information of filtering and smoothing distribution from
\code{dlm.fit} object}
\usage{
forecast(
  x,
  horizon,
  xreg = NULL,
  xreg_tf = NULL,
  prob_interval = c(0.05, 0.2),
  state_parameters = FALSE,
  discount = FALSE
)

forecast(
  x,
  horizon,
  xreg = NULL,
  xreg_tf = NULL,
  prob_interval = c(0.05, 0.2),
  state_parameters = FALSE,
  discount = FALSE
)
}
\arguments{
\item{x}{an object of class \code{dlm.fit}.}

\item{horizon}{integer. The forecast horizon.}

\item{xreg, xreg_tf}{matrix. The corresponding matrix with the predictors for the regression and transfer function components.}

\item{prob_interval}{vector with the probability of the credibility interval.}

\item{state_parameters}{logical. Should returned the future prior moments of state
parameters? Default is \code{FALSE}.}

\item{discount}{logical. Whether to use discounting when forecasting. Default is \code{FALSE}.}

\item{...}{currently not used.}
}
\value{
A \code{data.frame} with the results for each time.

If \code{state_parameters=FALSE} then return a \code{data.frame} with the response forecasts. Otherwise return a \code{list} with two components: \code{response} a data.frame with the response forecasts and \code{state} a data.frame with the state parameters forecasts.
}
\description{
Tidy the results of the forward filter or backward smoother into
a \code{data.frame}.
}
