% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\alias{forecast_marginal}
\alias{forecast_aR}
\alias{forecast.dlm}
\alias{forecast_marginal.dlm}
\alias{forecast_aR.dlm}
\title{Forecast Bayesian Dynamic Linear Models.}
\usage{
forecast(object, horizon, ...)

forecast_marginal(object, horizon, ...)

forecast_aR(object, horizon, ...)

\method{forecast}{dlm}(
  object,
  horizon,
  xreg = NULL,
  interval = TRUE,
  prob_interval = 0.1,
  state_parameters = FALSE,
  ...
)

\method{forecast_marginal}{dlm}(object, horizon, xreg = NULL, ...)

\method{forecast_aR}{dlm}(object, horizon, ...)
}
\arguments{
\item{object}{fitted model object of class \code{dlm.fit}.}

\item{horizon}{integer. Forecast horizon.}

\item{...}{currently not used.}

\item{xreg}{matrix. Future values of possible regressors.}

\item{interval}{logical. Should exact credible interval be returned?}

\item{prob_interval}{numeric. The probability level of the credible interval.}

\item{state_parameters}{logical. Should returned the future prior moments of state
parameters? Default is \code{FALSE}.}
}
\description{
Method \code{forecast} applied to objects of class \code{dlm.fit}
to perform marginal forecast of predictive and/or state distributions.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.

Migon, H. M. and Gamerman, D. (1993). Generalized exponential growth models: A Bayesian approach,  \emph{Journal of Forecasting}, \bold{12}, 573--584
}
\author{
André F. B. Menezes
}
