% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{get_daughter_states}
\alias{get_daughter_states}
\title{Get daughter states}
\usage{
get_daughter_states(node)
}
\arguments{
\item{node}{the node whose terminal daughter states should be obtained}
}
\value{
a vector with all states
}
\description{
Get all the states of all the terminal nodes under this node
}
