% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{get_internal_distance}
\alias{get_internal_distance}
\title{Get the internal distance}
\usage{
get_internal_distance(node)
}
\arguments{
\item{node}{the node}
}
\value{
the internal distance
}
\description{
Get the total distance between all nodes
}
