% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{get_output}
\alias{get_output}
\title{Get the output data frame}
\usage{
get_output(all_stats, shuffled_stats)
}
\arguments{
\item{all_stats}{the list containing all the values of all the stats for all the normal trees}

\item{shuffled_stats}{the list containing all the values of all the stats for all the randomized trees}
}
\value{
Undecided
}
\description{
Calculate the means, upper and lower confidence intervals for both the normal and randomized trees. Then calculate the significance.
}
