% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{get_parsimony}
\alias{get_parsimony}
\title{Get parsimony score}
\usage{
get_parsimony(node)
}
\arguments{
\item{node}{the node that's checked for parsimony}
}
\value{
the parsimony score
}
\description{
Get the parsimony score following Fitch's algorithm
}
