% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_phylogenetic_distance}
\alias{get_phylogenetic_distance}
\title{Get phylogenetic distance}
\usage{
get_phylogenetic_distance(node)
}
\arguments{
\item{node}{the node of which the phylogenetic distance is calculated}
}
\value{
the phylogenetic distance
}
\description{
Get the phylogenetic distance following Faith's algorithm. This distance is calculated like so:
 - 1 x the distance of all terminal nodes
 - 1 x the distance of all monophyletic internal nodes
 - 2 x the distance of all non-monophyletic internal nodes
}
