% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{get_possible_states}
\alias{get_possible_states}
\title{Get possible states}
\usage{
get_possible_states(xmldict, userinput)
}
\arguments{
\item{xmldict}{the xml dictionary}

\item{userinput}{the attribute the user marked as the state attribute}
}
\value{
a vector with all possible states for that attribute
}
\description{
Get tall the possible states for the state attribute
}
