% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{get_total_distance}
\alias{get_total_distance}
\title{Get the total distance}
\usage{
get_total_distance(node)
}
\arguments{
\item{node}{the node}
}
\value{
the total distance
}
\description{
Get the total distance of all the nodes
}
