% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{get_UniFrac}
\alias{get_UniFrac}
\title{Calculate UniFrac}
\usage{
get_UniFrac(node)
}
\arguments{
\item{node}{the tree whose AI should be calculated}
}
\value{
the association index
}
\description{
Calculate the UniFrac of the whole tree
}
