% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{highest_mono}
\alias{highest_mono}
\title{Get MC scores}
\usage{
highest_mono(treelist, possible_states)
}
\arguments{
\item{treelist}{a tree object}

\item{possible_states}{a vector of all possible states}
}
\value{
a list with the MC score for each monophyletic clade
}
\description{
Get the largest monophyletic clade for each state
}
