% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{make_internal_node}
\alias{make_internal_node}
\title{Make an internal node}
\usage{
make_internal_node(treesplit, pos, statesdict, state_attribute)
}
\arguments{
\item{pos}{the position in the string}

\item{statesdict}{the xml dictionary of the states}

\item{state_attribute}{the attribute that determines the state}

\item{tree}{the split string of the tree}
}
\value{
the internal node
}
\description{
Creates an internal node by making the intternal nodeds and terminal nodes in it
}
