% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{make_tree}
\alias{make_tree}
\title{Make the tree}
\usage{
make_tree(tree, statesdict, state_attribute)
}
\arguments{
\item{tree}{the string of the tree}

\item{statesdict}{the xml dictionary of the states}

\item{state_attribute}{the attribute that determines the state}

\item{pos}{the position in the string}
}
\value{
the tree
}
\description{
Creates the tree by making all the nodes and connecting them
}
