% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{market_share}
\alias{market_share}
\title{Market share}
\format{
A \code{\link{data.frame}} with RR observations and CC columns.
}
\source{
XXXX.
}
\usage{
data(market_share, package = "RBATS")
}
\description{
Market share
}
\author{
André Felipe Menezes
}
\keyword{datasets}
