% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{NTI_NRI}
\alias{NTI_NRI}
\title{Calculate NTI and NRI}
\usage{
NTI_NRI(utree, possible_states, userinput, xmldict)
}
\arguments{
\item{utree}{the tree in the ape phylo format}

\item{possible_states}{a vector with all possible states}

\item{userinput}{the state attribute}

\item{xmldict}{the xml dictionary}
}
\value{
the association index
}
\description{
Calculate both the Net Relatedness Index and the Nearest Taxa Index.
}
