% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dlm.fit}
\alias{plot.dlm.fit}
\title{Plot method for Bayesian Dynamic Linear Models.}
\usage{
\method{plot}{dlm.fit}(
  x,
  what = c("predictive", "posterior"),
  type = c("filter", "smooth", "both"),
  interval = type != "both",
  parm = NULL,
  ...
)
}
\arguments{
\item{x}{fitted model object of class \code{dlm.fit}.}

\item{what}{character. Should plot the \code{"predictive"} or \code{"posterior"}
state moments?. Default is \code{"predictive"}.}

\item{type}{character. Should plot the \code{"filter"}, \code{"smooth"}, or
\code{"both"}?. Default is \code{"filter"}.}

\item{interval}{logical. Should exact credible interval be plotted? Default is
\code{TRUE} for filter distribution and FALSE for \code{"smooth"} and \code{"both"}.}

\item{parm}{vector. a specification of which parameters are to be plotted, a vector of names. If missing, all parameters are considered.}

\item{...}{additional plotting arguments that affect the plot.}
}
\description{
Method \code{plot} applied to objects of class \code{dlm.fit}.
Plot historical data with forecasts and credible interval. Also, plot the filter and
smooth state parameters.
}
\author{
André F. B. Menezes
}
