% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print_start}
\alias{print_start}
\title{Starting message}
\usage{
print_start(treefile, apetrees, possible_states, state_attribute, reps)
}
\arguments{
\item{treefile}{the string of the path to the .trees file}

\item{apetrees}{the raw read trees from the .trees file}

\item{possible_states}{a vector of all states for the given state attribute}

\item{state_attribute}{the state attribute}

\item{reps}{the number indicating the amount of randomized trees that should be made}
}
\description{
Print a message before starting the analyzing process to inform the user of the input that is used
}
