% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{process_xml}
\alias{process_xml}
\title{Process the xml file}
\usage{
process_xml(input_file)
}
\arguments{
\item{inputfile}{the xml file}
}
\value{
the xml dictionary
}
\description{
Processes the xml file to create the xml dictionary
}
