% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBATS-package.R
\docType{package}
\name{RBATS-package}
\alias{RBATS-package}
\title{Overview of the \pkg{RBATS} package}
\description{
The \pkg{RBATS} R package provides functionality to work
with Bayesian Forecasting and Dynamic Models.
}
\author{
André F. B. Menezes \email{andrefelipemaringa@gmail.com}
}
