% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrospective.R
\name{retrospective_moments}
\alias{retrospective_moments}
\alias{retrospective_moments.dlm}
\title{Smoothing moments of state parameters}
\usage{
retrospective_moments(object, time)

\method{retrospective_moments}{dlm}(object, time)
}
\arguments{
\item{object}{an object of \code{dlm}.}

\item{time}{integer. The time index.}
}
\description{
Method to smoothing moments of state parameters and the
one-step ahead predictive distributions for objects of class  \code{dlm}.
}
\references{
Prado, R.; West, M. Time Series Modeling, Computation, and Inference. CRC Press, 2010.
}
\author{
André F. B. Menezes
}
