% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{set_monoweights}
\alias{set_monoweights}
\title{Finish internal node}
\usage{
set_monoweights(thisnode)
}
\arguments{
\item{thisnode}{the internal node that should be finished}
}
\value{
the finished internal node
}
\description{
Finishes the internal node by setting the states, state weights, and whether or not the node is monophyletic.
}
