% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{set_state_attribute}
\alias{set_state_attribute}
\title{Set the state attribute}
\usage{
set_state_attribute(xmldict, userinput)
}
\arguments{
\item{node}{the tree whose AI should be calculated}
}
\value{
the association index
}
\description{
Calculate the UniFrac of the whole tree
}
