% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{shuffle}
\alias{shuffle}
\title{Shuffler}
\usage{
shuffle(xml, userinput)
}
\arguments{
\item{xml}{the xml dictionary}

\item{userinput}{the state attribute that should be shuffled}
}
\value{
the new shuffled xml dictionary
}
\description{
Shuffle the states.
}
