% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{smart_pick}
\alias{smart_pick}
\title{Pick a good state attribute}
\usage{
smart_pick(xmldict)
}
\arguments{
\item{xmldict}{the xml dictionary}
}
\value{
a good or the first attribute
}
\description{
Pick an attribute that can be used as a state attribute. Returns either a good state attribute or the first attribute.
}
