% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{telephone_calls}
\alias{telephone_calls}
\title{Telephone calls}
\format{
A \code{\link{data.frame}} with RR observations and CC columns.
}
\source{
XXXX.
}
\usage{
data(telephone_calls, package = "RBATS")
}
\description{
Telephone calls
}
\author{
André Felipe Menezes
}
\keyword{datasets}
