% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempcon.R
\name{topmono}
\alias{topmono}
\title{Get streaks}
\usage{
topmono(node, statelist)
}
\arguments{
\item{node}{the node of which the monophyletic clades are calculated}

\item{statelist}{a list with a vector for every state}
}
\value{
the statelist with all the monophyletic clades
}
\description{
Make a list with every monophyletic clade for each state
}
