% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dgegm.R
\name{update.dgegm}
\alias{update.dgegm}
\title{Update the moments of state parameters}
\usage{
\method{update}{dgegm}(object, y, ...)
}
\arguments{
\item{object}{an object of \code{dgegm} class.}

\item{y}{vector. Observed value of time series.}

\item{...}{currently not used.}
}
\description{
Methods to update the moments of state parameters and compute the
one-step ahead predictive distributions for objects of class  \code{dlm} and
\code{dgegm}.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.

Migon, H. M. and Gamerman, D. (1993). Generalized exponential growth models: A Bayesian approach,  \emph{Journal of Forecasting}, \bold{12}, 573--584
}
\author{
André F. B. Menezes
}
