% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dlm.R
\name{update.dlm}
\alias{update.dlm}
\title{Update the moments of state parameters}
\usage{
\method{update}{dlm}(object, y, X = NULL, ...)
}
\arguments{
\item{object}{an object of \code{dlm} or \code{dgegm}.}

\item{y}{vector. Observed value of time series.}

\item{X}{matrix. Regressors.}

\item{...}{currently not used.}
}
\description{
Methods to update the moments of state parameters and compute the
one-step ahead predictive distributions for objects of class \code{dlm} and
\code{dgegm}.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\author{
André F. B. Menezes
}
