% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_moments}
\alias{update_moments}
\alias{update_moments.dlm}
\alias{update_moments.dgegm}
\title{Update the moments of state parameters}
\usage{
update_moments(object, time)

\method{update_moments}{dlm}(object, time)

\method{update_moments}{dgegm}(object, time)
}
\arguments{
\item{object}{an object of \code{dlm} or \code{dgegm}.}

\item{time}{integer. The time index to update the moments.}
}
\description{
Methods to update the moments of state parameters and compute the
one-step ahead predictive distributions for objects of class  \code{dlm} and
\code{dgegm}.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.

Migon, H. M. and Gamerman, D. (1993). Generalized exponential growth models: A Bayesian approach,  \emph{Journal of Forecasting}, \bold{12}, 573--584
}
\author{
André F. B. Menezes
}
