% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{us_retail_employment}
\alias{us_retail_employment}
\title{US monthly retail employment}
\format{
A \code{\link{data.frame}} with 357 observations and 2 columns.
}
\source{
According to \code{fpp3} package: U.S. Bureau of Labor Statistics.
}
\usage{
data(us_retail_employment, package = "RBATS")
}
\description{
Data extract from \code{us_employment} of \pkg{fpp3} package.
It contains US retail employment data from January 1990 to June 2019.
}
\author{
André Felipe Menezes
}
\keyword{datasets}
