% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{vic_electricity_daily}
\alias{vic_electricity_daily}
\title{Daily electricity demand of Victoria, Australia}
\format{
A \code{\link{data.frame}} with 367 observations and 5 columns:
\itemize{
\item \code{time}: Date. The correspond day.
\item \code{demand}: Numeric. Total electricity demand per 1000.
\item \code{max_temperature}: Numeric. The maximum temperature in that day.
\item \code{holiday}: Logical. Indicator for if that day is a public holiday.
\item \code{day_type}: Factor. The type of day, it could be \code{holiday}, \code{weekday}, and \code{weekend}.
}
}
\source{
According to \code{tsibbledata} package: Australian Energy Market Operator.
}
\usage{
data(us_retail_employment, package = "RBATS")
}
\description{
Data extract from \code{vic_elec} of \pkg{tsibbledata} package.
It contains the daily electricity demand per 1000 and maximum temperature in Victoria, Australia.
for 2014.
}
\author{
André Felipe Menezes
}
\keyword{datasets}
