% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlm.R
\name{dlm}
\alias{dlm}
\title{Bayesian Dynamic Linear Model}
\usage{
dlm(
  polynomial = list(order = 1L, discount_factor = 0.95),
  seasonal = list(type = c("none", "free", "fourier"), period = NULL, harmonics = NULL,
    discount_factor = 0.98),
  regressor = list(xreg = NULL, discount_factor = 0.99),
  autoregressive = list(order = NULL, discount_factor = 0.998),
  transfer_function = list(order = NULL, xreg = NULL, discount_factor = 0.998),
  df_variance = 1,
  variance_law = list(type = "identity", power = 1),
  monitor = list(execute = FALSE, verbose = TRUE, start_time = 10L, bilateral = FALSE,
    bf_threshold = 0.135, location_shift = 4, scale_shift = 1, discount_factors =
    list(polynomial = 0.2, seasonal = 0.8, regressors = 0.8))
)
}
\arguments{
\item{polynomial}{list. Order of polynomial model.}

\item{seasonal}{list. Components to specify seasonal components.}

\item{regressor}{list.}

\item{df_variance}{numeric. Discount factor for observation variance. Use a beta-gamma random walk.}

\item{variance_law}{list. Variance law \code{type} and \code{power} parameter.}

\item{monitor}{list.
The variance law \code{type} are \code{identity}, \code{poisson}, \code{binomial},
and \code{power}. The variance law \code{power} should be numeric \eqn{p \geq 1}.}
}
\description{
Create object of class \code{dlm}.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\author{
André F. B. Menezes
}
