% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_dlm.R
\name{fit}
\alias{fit}
\alias{forward_filter}
\alias{backward_smoother}
\alias{fit.dlm}
\alias{forward_filter.dlm}
\alias{backward_smoother.dlm}
\title{Forward filtering and Backward Smoothing for Bayesian Dynamic Linear Models}
\usage{
fit(model, y, ...)

forward_filter(model, y, ...)

backward_smoother(model, filtering_parameters, ...)

\method{fit}{dlm}(model, y, m0, C0, n = 1, s = 1, smooth = TRUE, ...)

\method{forward_filter}{dlm}(model, y, m0, C0, n = 1, s = 1, ...)

\method{backward_smoother}{dlm}(model, filtering_parameters, ...)
}
\arguments{
\item{model}{a model object of class \code{dlm}.}

\item{y}{vector. Observed value of time series.}

\item{...}{currently not used.}

\item{filtering_parameters}{list. Historical prior moments
(\code{a} and \code{R}) and historical posterior moments
(\code{m} and \code{C}).}

\item{m0, C0}{vector and matrix. posterior mean vector and covariance matrix at time t-1.}

\item{n, s}{numeric. Prior sample size and mean for the variance, respectively.
For both parameter the default is 1.}

\item{smooth}{logical. Should smoothing be performed? Default is \code{TRUE}.}
}
\description{
Methods \code{forward_filter} and \code{backward_smoother}
for objects of class \code{dlm}. The implementation is in \code{C++}.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.

Prado, R.; West, M. Time Series Modeling, Computation, and Inference. CRC Press, 2010.
}
\author{
André F. B. Menezes
}
