% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\alias{forecast.dlm.fit}
\alias{forecast_marginal.dlm.fit}
\alias{forecast_aR.dlm.fit}
\title{Forecast Bayesian Dynamic Linear Models.}
\usage{
forecast(object, ...)

\method{forecast}{dlm.fit}(
  object,
  xreg = NULL,
  horizon,
  interval = TRUE,
  level = 0.1,
  state_parameters = FALSE,
  ...
)

forecast_marginal.dlm.fit(object, xreg = NULL, horizon)

forecast_aR.dlm.fit(object, horizon)
}
\arguments{
\item{object}{fitted model object of class \code{dlm.fit}.}

\item{...}{currently not used.}

\item{xreg}{matrix. Future values of possible regressors.}

\item{horizon}{integer. Forecast horizon.}

\item{interval}{logical. Should exact credible interval be returned?}

\item{level}{numeric. The probability level of the credible interval.}

\item{state_parameters}{logical. Should returned the future prior moments of state
parameters? Default is \code{FALSE}.}
}
\description{
Method \code{forecast} applied to objects of class \code{dlm.fit}
to perform marginal forecast of predictive and/or state distributions.
}
\references{
West, M.; Harrison, J. Bayesian Forecasting and Dynamic Models. Springer, 1997.
}
\author{
André F. B. Menezes
}
